requirements.txt
requests>=2.28.1
pytest>=7.2.0
allure-pytest==2.12.0


curlify~=2.2.1
assertpy>=1.1
dynaconf~=3.1.2
urllib3~=2.0.6
singleton~=0.1.0
Faker~=19.10.0
marshmallow~=3.20.1